+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next
 
-- Adds additional options to TextBlock component
+- Add additional options to TextBlock component
 - Add "Return to top" component
 - Add `dollarFormat` utility function
 - Add `kebabCase` utility function